fix: admin RLS policies using is_admin() function to prevent recursion
The 999_enable_rls_policies.sql migration was overwriting the admin policies
that were fixed in 003_fix_admin_rls_recursion.sql with recursive versions.

This caused admins to get 'You don't have permission to access this' errors
when trying to approve/reject helpers because the RLS policy was causing
infinite recursion when checking if the user is an admin.

This migration fixes all admin policies to use the is_admin() SECURITY DEFINER
function which properly bypasses RLS recursion.